Updated openstack/openstack Project: openstack/tripleo-heat-templates  ff5b7a1d35bbebe110f424b10d213facfca21dff
Modify external loadbalancer environments to use new FixedIPs

This change modifes the environment files
external_loadbalancer_vip.yaml and external_loadbalancer_vip_v6.yaml
to take advantage of the new mechanism for setting FixedIPs for each
VIP. The previous mechanism was used to set the VIP addresses
directly, but each VIP now contains a FixedIPs parameter which sets
the IP for the VIP when it is created. The advantage of the new
mechanism is that we no longer have to set each VIP to noop.yaml in
the resource registry, since vip.yaml can now handle both automatic
and fixed IP selection.

Change-Id: I6e66ea6041fe7357c01b8f79f04e2533579206dc