Updated openstack/openstack Project: openstack/python-zaqarclient  73da2fcce2340294ddf5aaaaafe9526706a2c65a
translate all command help strings

Leverage the new cliff command class attribute (_description)
to get the help of a command, this allows us to mark strings
for translation[1]. We could not do this before since the help
was grabbed from the docstring.

The old help strings are still left to keep backward compatible.

[1]:https://review.openstack.org/#/c/397321/

Co-Authored-By: wanghao <sxmatch1986@gmail.com>

Change-Id: If9c6009eee4dbee87deda8aba63db088ced4f014