Updated openstack/openstack Project: openstack/nova  42ebedfcde1bb45cc57a949632dfe583cf7aaa66
nova-status: check for cells v2 upgrade readiness

This adds the nova-status upgrade readiness check for cells v2
which is very similar to the 030 API DB schema migration except
we only check the cell and host mappings, we don't rely on flavors
being in the database or not. Since this check requires that you
run 'nova-manage cell_v2 simple_cell_setup' first, or a manual
variant thereof, we really only care about the cells v2 parts.

Part of blueprint cells-scheduling-interaction

Change-Id: Id52c0b7373585286292d252236bfbba13e13fc69