Updated openstack/openstack Project: openstack/sahara  e1602b7521dc91ce3b18087b006e4c37ebbd4165
Adding tenant_id to regex_search

In order to allow the creation of clusters templates and node group
templates with the same across projects, even when set to public, we need to
filter only from the requester's project when creating a new resources.
This will narrow the name check to same project only.

Change-Id: Icc39c7e680c2fff583069a5522b2a11afb3e239b
Closes-bug: #1584680