Updated openstack/openstack Project: openstack/python-novaclient  aeec4d11b2dae769bfe020dcd95ec607afaeed55
Fix "Message object do not support addition".

In the 7.0.0 Novaclient release we added some warning
log about user should using "endpoint_type" rather than
"interface" when init novaclient:
https://github.com/openstack/python-novaclient/blob/master/novaclient/client.py#L312-L327

https://github.com/openstack/python-novaclient/blob/master/novaclient/client.py#L249-L272

This is now causing a lot jenkins failures acroos projects
that is using novaclient with enable_lazy set to False.
As in this kind of scenario the warning message is a message
object instead of unicode sting and it cannot be added:
http://git.openstack.org/cgit/openstack/oslo.i18n/tree/oslo_i18n/_message.py#n227

and "Message object do not support addition" will raise.

Related Jenkins Error:
Ceilometer:
http://logs.openstack.org/29/333129/3/check/gate-ceilometer-python27-ubuntu-xenial/bd125e7/

Closes-bug: #1652414
Change-Id: I11a490f759fdac9707c1321c9659da2605196a94