Updated openstack/openstack Project: openstack/keystone  6a7eea1574dc852981d9b9f4512be1a0ebadd443
Remove unused variables from federation tests

A test method in test_v3_federation.py didn't do anything with the
output it received because it was expecting the call to return a
failure status code, which is check with `expected_status`. This
commit just removes the unused variables since they aren't
asserting anything about the test behavior.

Change-Id: I5eb374f8504a2818384d2631f905dc323d548f60