Updated openstack/openstack Project: openstack/neutron  a2f41687e31064b1ee13492f6c8a7d485b545db1
Pecan: Change Quotas to use original controllers

Previously, pecan was using custom controllers to handle quotas API
calls.  This resulted in the pecan custom controllers becoming out of
date because it was only changed in the original controllers.  This
patch removes the use of the pecan controllers and uses shim controllers
to redirect to the original quota controllers.

Change-Id: If673216db51439b9b404d9f18ee5dbf7228c8d2a