Updated openstack/openstack Project: openstack/nova  7391593f3f49b5fffa5c8a8b033fcbc52d168f2a
Ignore 404s when deleting allocation records

We don't need to emit a warning when trying to delete
an allocation record that doesn't exist. We see these
warnings all over the CI runs because of Tempest tests
that delete a server while it's building, in which case
the allocation record may have never been created.

Change-Id: I472fe1a091dc2524c50873d4fd916bcc535f51c7
Closes-Bug: #1652155