Updated openstack/openstack Project: openstack/tempest  1b8f6f00eedf957990d47bc423a48f0760adfb79
Fix container cleanup in test_account_bulk

BulkTests needs to cleanup containers after each tests, and
it tries to so by setting an instance attribute, a list of
containers, and using a class method delete_containers,
which uses a class attribute to discover containers to be
deleted.

Fixing this by making delete_containers a static method
which accepts the list of containers and an input.

Partial-bug: #1650205
Partial-bug: #1609156
Change-Id: I3651c2e06ca52b64e43e707de9af7d45062629bc