Updated openstack/openstack Project: openstack/cinder  ffb5eef8c84d18b3a77d44e0623f6bc36f5d21cf
Update replication property in capabilities

Our current standard capabilities are reporting the nonexistent
"replication" property to the scheduler, and this is then exposed to the
users via the "get_capabilities" REST API.

Currently users will see a "replication" property defined as:

  {u'type': u'boolean',
   u'description': u'Enables replication.',
   u'title': u'Replication'}

And while the definition is correct, the name of the property isn't,
since replication v2.1 uses "replication_enabled" instead, which defeats
the whole purpose of reporting the standard properties.

This patch updates the reported name so it matches existing
implementation.

Closes-Bug: #1642403
Change-Id: Ia39c24d1b256fb3b1d1ee2f1a8c47a04dbb1c3b7