Updated openstack/openstack Project: openstack/heat-translator  0563989c56b092071332bbf35034cebdff606358
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove heat-translator.

Change-Id: I3117119a67315c513fbd689eb23cda22c1af4ae1