Updated openstack/openstack Project: openstack/neutron  4f8a106cb504016a2f5823b00233ee4c5498ba7e
Change passing session to context in segments db functions

We have a number of functions that expect to get session as one
of the arguments. Passing session is not correct and this prevents
using new enginefacade which expects context to be passed and
session will be injected in current context.

NeutronLibImpact

Partially-Implements blueprint: enginefacade-switch

Change-Id: Ie1c54138ceaf9ac6f0674ad2786d4aaea9c80f73