Updated openstack/openstack Project: openstack/trove  34b1538af861693a4a311404e6ccad1ef87848a8
get rid of a couple of old oslo-incubator object references

Found these while looking around the rpc message code; these paths are
no longer valid. Didn't want to combine this change with the one to
add signing/verification as it appears to be more of a cleanup thing.

TRANSPORT_ALIASES is no longer used so I've got rid of that as well.

Change-Id: I398f1286bcb08740a6125d4e773c9de6708e1d43