Updated openstack/openstack Project: openstack/trove  998c021d36edcc86dd2b62e94522be7759e6c253
trove pylint cleanup(s)

This commit includes two cleanups in trove-pylint.

First, the deprecation warning is because the text.ParseableOutputFormat is
being deprecated[1] and produces a deprecation warning as shown in the
bug.

Second, while looking into 1650816, I stumbled upon 1650888 which is
also fixed. A block of code was wrongly indented and when I disabled
"-E" to generate warnings, produced a crash in pylint. If the line
generated doesn't match the regular expression, the various parameters
to the next call (to ignore()) are not set and you get unbound
variable errors.

[1] https://github.com/PyCQA/pylint/blob/master/pylint/reporters/text.py#L163

Change-Id: I43bd6569a98eeb1b326e4a0ca1addaf97c0830a5
Closes-Bug: #1650530
Closes-But: #1650888