Updated openstack/openstack Project: openstack/ironic  74a052858eee0a65f1e8b3f21bdd74d099796e5b
Auto-set nullable notification payload fields when needed

Currently, the populate_schema method in the notification payload base
class raises an exception whenever any payload field in the notification
schema is missing from the populating object, even if that field is
nullable in the payload.

This behavior is not expected and results in unintended failures when
populating a schema with an object missing nullable payload fields.

Change-Id: I05d7f373ee441b63d833b3c90cc5f225cd81b5be
Closes-Bug: #1651517