Updated openstack/openstack Project: openstack/python-manilaclient  055acfd5f034839851f62767a44db1b7bc543dc1
Add Constraints support

Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove python-manilaclient.

Change-Id: I2d5bb052c7e499175ef3d32fdc35c9e67dc260ac