Updated openstack/openstack Project: openstack/ironic  831f86d8b919e866b975c18233f65f9302e07889
Fix wrong exception message when deploy failed

Currently, if conductor gets an exception while trying to deploy the
node, we will get an exception message like:

Asynchronous exception for node $UUID: Node failed to get image for
deploy. Exception: Failed to continue agent deployment.

It's not right because the conductor may successfully fetch the image
and fail in other reasons, like fail to connect to neutron.

Change-Id: Ide4bb749365709a3ec321b74a58abdc0006e34b8