Updated openstack/openstack Project: openstack/trove  46f07e5a2fc41a0b2cf0d803f8f17a7ab43e37a6
Fix trovestack clean to remove nova instances

If you run 'trovestack clean' any nova instances are not deleted. This
seems to be caused by the fact that the OS_PROJECT_NAME env variable
does not override the value found in the clouds.yaml file.

This was fixed by having the value passed in as --os-cloud change
based on the project name being cleaned.

Change-Id: I7d360b14bcacf3617bbddd3cd68a341a369bb40a
Closes-Bug: 1653435
Co-Authored-By: Peter Stachowski <peter@tesora.com>
Co-Authored-By: Doug Shelley <doug@tesora.com>