Updated openstack/openstack Project: openstack/keystone  068b1df8415af533b3abd703bfddb151c946fd9c
Fix cloud_admin rule and ensure only project tokens can be cloud admin

The current rule fails to load with oslo.policy, the correct
value used to determine the admin project for the cloud_admin should
simply be: `is_admin_project:True`, since that is what is stored
in oslo.context.

This problem was masking a more serious issue that domain admin tokens
could be misinterpreted as cloud admin tokens.

Change-Id: I3ea562c01e06e6c519fdaec3ab6e1dac204ced71
Closes-Bug: 1547684
Closes-Bug: 1651989