Updated openstack/openstack Project: openstack/barbican-specs  60eded4f12f57293ea822d6b667641511c460573
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/25ec3803cb84c530b5d0bf7f44d01077

Change-Id: I47804903a4dbeebe69f406e1cef0709e91e30d00