Updated openstack/openstack Project: openstack/castellan  6fbf687fc44f3c570b42f4938df40407702c877e
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/b532dc00916feefa9e94fe5686495095

Change-Id: I27e9461332f8538673995b4ffe7a84349e249382