Updated openstack/openstack Project: openstack/barbican  ff4472157f01a0dd822f946a5e00dde8ee09635c
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/b90a92195027415811203758f32da4be

Change-Id: Id0ab11a8a25ffadd6b7425a172f7c32c5e9f86bf