Updated openstack/openstack Project: openstack-infra/storyboard-webclient  28b480942521ae5fccdab64b48adbee510048dd4
Allow teams to be added to the permission list for private stories

This commit adds a way to add teams to the permission list for private
stories. This makes it easy to use for stories that should only be
visible to a given set of people; the members of that set can now be
managed by the teams UI rather than on a per-story basis.

Change-Id: I7d3adde5d6cf9b4d5b7e63df06d2266f7b824666
Depends-On: I16a7451e8c24da352357abb0590d0ace5b4e77f4