Updated openstack/openstack Project: openstack/horizon  c789a7aba5cebe8bd54dfa0c6b602ef18fe6d43f
Specify POLICY_CHECK_FUNCTION as a string

We don't want code in our settings.py and local_settings.py, and in
particular we don't want to have to import Python objects from all over
to set them as setting values -- instead, we can specify those as import
path strings. This also solves problems with importing order and loops.

This change is backwards-compatible, in the sense that you can still
import the objects directly and set them as the setting values.

Partially-Implements: blueprint ini-based-configuration
Change-Id: I8a346e55bb98e4e22e0c14a614c45d493d20feb4