Updated openstack/openstack Project: openstack/cliff  20319083f2c7eba6913fe282ad82cb74c0e15064
Fix command order

When there are two commands: one 'foo', that expects a positional
argument, and a second command 'foo bar'.
Trying to enter the command 'foo bar' will invoke the command
'foo' with position argument 'bar', and not the second command
'foo bar'.
This is due to the fact that the find command method first
checks for the first word as a command.
By trying first to match all arguments this problem is solved.

Closes-Bug: #1618673
Change-Id: I24ea39813ad30ec095ea18baad3a7b088fa8108f