Updated openstack/openstack Project: openstack/os-net-config  d929214c7a904da57a406dd56ac06f001fdc7d7b
Add support for enabling hotplug on interfaces

This patch adds support for enabling hotplugging on interfaces (disabled
by default). This is useful for configuring SR-IOV root devices so that
they "return" to the system when no longer used by a VM.

Note: also updates an invalid value in the interface and ib_interface
sample files.

Partial-Bug: #1639901
Change-Id: Idfc17d6f20bb306271838895bc53f4b109dd664d