Updated openstack/openstack Project: openstack/keystone  d5ce8ea0ed6a3b9f748427d1b552fb88ad738545
Add --check to keystone-manage db_sync command

This patch adds a new command to the db_sync upgrade commands. --check
will check the current state of the users upgrade repos and relay info
back to the user based on what version each command is currently at and
if the user has any outstanding db_sync commands left to run. It will
also notify the user if the db_sync commands were not upgraded in order

Closes-Bug: 1642212
Change-Id: I79d3640a780d624f14059fe311fafa0542c03357