Updated openstack/openstack Project: openstack/nova  e8996713f859e5e5deba236ff52543ee2e616627
Make servers api view load instance fault from proper cell

Right now with CellsV2 the servers api will fail to load the fault
for an instance because it does not target it to a particular cell.
This makes us look up the instance mapping and use it to target
loading of the fault.

Change-Id: Ic2df2c3cfa66a91bd20437f534287e770b85f51d