Updated openstack/openstack Project: openstack/cinder  fdb72af62f421a639cc490373c82cd22484e948d
Set replication_status on volume creation

The scheduler is currently always setting `replication_status` to
"disabled" regardles of the real replication status.

Some drivers are updating it on volume creation, but even in that case
the user will have received the API REST response with the
`replication_status` set to "disabled" and will have to do a `show` of
the volume once the creation has been completed to confirm if the volume
is actually replicated or not.

This patch changes this and allows the scheduler to correctly set the
`replication_status` on volume creation based on the standard property
`replication_enable`.

This doesn't solve 100% of the cases, as a Cloud Administrator may not
be using the `replication_enable` extra specs in volume types to set
replication because it's using a backed that is replicating at
backend/pool level.

In that case there will be a discrepancy, just like we have now, until
the driver completes the creation and returns the model update for the
volume.

Closes-Bug: #1643883
Change-Id: Id4df2b7ad55e9b5def91329f5437da9caa185c30