Updated openstack/openstack Project: openstack/stevedore  57e4c9cde4022cc0f50fe636e5f319abd658f4dd
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

b'https://gist.github.com/736af151582e8886653d8de583c44810\n'

Change-Id: I03cf703c02cadfe3b21bd7b6b35619fd30bb4d17