Updated openstack/openstack Project: openstack/neutron  7c771abe0e7ca51bc5a5efecb6bea5e85129891c
Account for unwire failures during OVS trunk rewiring operations

If a failure occurs while unwiring stale ports, we currently
ignore the outcome of the operation, whereas we should at least
warn the user that something is not quite as it is supposed
to be.

This patch makes sure that trunk status is accurately reflected
after a sequence of unwire+wire operations for a given trunk.

Closes-bug: #1649887

Change-Id: I3b6ed57e00c0146babe23ea6ed0ca14e83020d26