Updated openstack/openstack Project: openstack/cinder  9f00487d791e549aa211562c780452469b1ba467
Fix python integer interpretation in Py2 and Py3

If user passes large number "11111111111111111111111111111" as size to
the create volume api, then the api behaves inconsistently depending
on whether the cinder-api service is running on Py2 and Py3.
In case of Py2, it returns following error message:

ERROR: Invalid input received: Volume size '11111111111111111111111111111'
must be an integer and greater than 0 (HTTP 400) (Request-ID: req-abe6cd5
e-a0c6-4e0d-ba05-c9eceef547bf)

But in case of Py3, it passes this validation.

This patch fixes this inconsistency issue by checking isinstance(size, int)
to isinstance(size, six.integer_types).

NOTE:
The unit test written in the patch will fail on master on Py2 environment
and it's written to just ensure that the issue is resolved for Py2.

Closes-Bug: #1651103
Change-Id: Ib43b96d458f0d4e9477886bd19c35a7e05664929