Updated openstack/openstack Project: openstack/nova  12f65afd9d7a8ef80ab411633cf3ba2279c95a54
Make test_security_groups work with CONF.use_neutron=True by default

The tests in this module are all based on using nova-network for
security groups so we need to explicitly set use_neutron=False for
these tests. The security groups REST API is tested using Neutron
in the test_neutron_security_groups module.

Because of the joy of multiple inheritance we do have to modify
the test_neutron_security_groups tests to set an attribute that is
used in the base test classes before creating the controller object.

Part of blueprint use-neutron-by-default

Change-Id: I814949630bcef296773c14bff21b921d58c25d8b