Updated openstack/openstack Project: openstack/nova  c061f29abc698d33a8aba3bc787c8b99e2108e37
Make test_tenant_networks pass with CONF.use_neutron=True by default

os-tenant-networks only supports listing networks and showing details
about a given network when using Neutron. Creating a network through
this API when using Neutron result in a 503 response, and trying to
delete one will result in a 500 response because the NotImplementedError
isn't handled in the delete() method (maybe it should be, but that's not
going to be changed in this patch).

This breaks the tenant network tests into two classes and handles the
conditional logic based on if neutron is being used or not, and also
fixes the mocks for the index/show methods so they more generically
mocking the network API in the controller object rather than nova-network
globally.

Part of blueprint use-neutron-by-default

Change-Id: I4acccb4de1a4c6481866cabd828cdd63bf9381d0