Updated openstack/openstack Project: openstack/nova  190d75927456809d169d1c1875a7314831152a39
Changed NUMACell to InstanceNUMACell in test_stats.py

Tested methods consume_requests() and consume_requests() operate
with InstanceNUMACell objects but because of mistake NUMACell
objects were used instead. It may cause the problem if these
methods will try to use some field which is not a field of NUMACell
but is a field of InstanceNUMACell.

Also some indents were removed from assertTrue/False

Change-Id: Ib41c9b22bfd22e769fac729f09d95d6bc8ba8e6b