Updated openstack/openstack Project: openstack/python-openstackclient  5750ff920eb4c6af5c575baa9e4308fd7676d632
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/30af7c8ad287756d38d131d1f5dceae0

Change-Id: I2d906fa0675463e21f2786d93f464c465f2bc3b1