Updated openstack/openstack Project: openstack/horizon  2694190df40df738eaa090922fc6c56b37aeb8f0
Remove duplicated modal_header statements

Across the codebase we've been using modal_header values that are
identical (or very close to) the page_title values. This patch sets the
modal_header to default to the page_title value, and cleans up a few
inconsistencies. Also deleted a couple of redundant submit_label values
that were identical to the default ("Submit").

Change-Id: I88815c3801c29b3fbc41e0cb426a50653255595f
Closes-Bug: 1653960