Updated openstack/openstack Project: openstack/nova  1bcf3b553ae8665dc6308d1bd11898efb75d3a41
Do not post allocations that are zero

When posting allocations for VMs with a non-local disk, the code was
sending an allocation of DISK_GB = 0, which violated the min_unit
resource constraint. Since it is pointless to enter an allocation of
zero units, this patch removes any items in the generated allocations
dict if the value is zero.

Closes-Bug: #1647316

Change-Id: I59eff4310e67ceb74086f5ade1a637f8cccec7ed