Updated openstack/openstack Project: openstack/cliff  1dd3edafab1a34d194c4510310653ad7e2cbb582
Remove tox environments that no longer work

Remove neutronclient-stable and openstackclient-stable which are designed to
run the latest cliff code/change under review against those 2 libraries to
reduce the chances of regressions.  At some point in the past they ceased
to function as expected and rather then correct them remove them as they're
not used in the gate anyway.

Change-Id: I8fa88134c730ed55afc40c6513f135912cbed5e6