Updated openstack/openstack Project: openstack-dev/devstack  85879f1f8157a60fc190eb2731e4edd0deb1dae3
Use the installed swift scripts

This commit switches how scripts we use to launch the installed version
in the path. Previously the scripts were manually executed in the source
repo, but this has issues if you're trying to run with py3 in a system
where python == py2. Setuptools already does the shebang magic for us
at install time, so we just need to use the installed version of the
script.

Change-Id: Iaa4d80ec607a2aa200400330e16cad3a4ca782ac