Updated openstack/openstack Project: openstack-infra/project-config  31c5cdcba4b4bfd79f59bbe6074334fd76b32f1c
Use gate_hook for Neutron rally scenario

Neutron rally scenario on master requires non-default
extension. (trunk)  The gate hook is planned to be
used to enable it.

Note: the second argument (scenario) is currently unused
but added for future possible use.

This is the second step of the following plan:

    1. make neutron's gate_hook accept "rally" but do nothing.
    2. make project-config change to consume the gate_hook
    3. make the rest of neutron change

Partial-Bug: #1643451
Change-Id: I574fc28ab010f9ba77e257ffa436054a999e8d91