Updated openstack/openstack Project: openstack-infra/release-tools  c99f81423d9c066faa452ace331b0daf073ba46d
adjust the emotions in the template

Not all of the adjectives we had in the emotion list are appropriate,
and some produce awkward phrasing when used in the structure "are $word
to". Adjust the emotion list to remove the items that fit poorly and
adjust the template and remaining items to use "We $emotion announce" as
the structure, allowing more variation in the form of emotion being
plugged in.

Change-Id: I976fc6cd4305edc8ea04629bd3e4d4e68ad6bbc7
Signed-off-by: Doug Hellmann <doug@doughellmann.com>