Updated openstack/openstack Project: openstack-infra/project-config  3a83266c082178d527bb92508b640044ee2ca007
neutron: add tripleo-ovb-ha in experimental

We're trying to debug a critical problem in Neutron that cause TripleO
HA deployments failing on pinging a floating-IP. This bug probably
affects all HA deployments in OpenStack.

To help in debug, add the ovb-ha job in experimental pipeline, so
Neutron developers can fix the problem and make sure it works in
tripleo.

Change-Id: I3a47092357c79486d33d9c155dd50f5febbb98fa