Updated openstack/openstack Project: openstack/api-wg  6ccd565824bfc1fc0e22c547c0ab65e340e6bdb8
Trivial fix: correct names of capitalization styles

The guideline on how to capitalize complex names uses the corresponding
style when mentioning the different styles to illustrate what each style
looks like. But the entry for 'MixedCase' is wrong. First, it is written
in CamelCase, but more importantly, "mixed case" refers to a variety of
styles of capitalization [0]. The only mixed case besides camelCase for
joined words is commonly called 'StUdLyCaPs' [1], in which the
capitalization is randomly altered.

[0] https://en.wikipedia.org/wiki/Capitalization
[1] https://en.wikipedia.org/wiki/Studly_caps

Change-Id: I41691bed53d5e30b445d042efbf6a7b5dce43185