Updated openstack/openstack Project: openstack/glance  922a2d72ce73cecd6faff9916b7bc54b903dac42
Image signature documentation modify barbican auth_endpoint

Add a configuration section to the image signature documentation
describing how to change the keystone endpoint of barbican if
keystone is not on localhost.

If auth_endpoint is incorrect then image signing will fail as barbican
will not be able to connect to keystone.

DocImpact

Added an example of changing the auth_endpoint.

Closes-Bug: 1620539

Change-Id: I8f8978f0c0dfc337e917c4d710dc8db1351070d2